Some updates for non-standard filenames, and ancient awful info.plists

+++ main.py
@@ -6,6 +6,7 @@ import time
 import re
 from shutil import copyfile
 from bs4 import BeautifulSoup
+from biplist import *
 
 # TIME BETWEEN UPLOADS (SECONDS)
 waittime = 180
@@ -49,35 +50,66 @@ if os.path.exists("Payload"):
 
 for file in filelist:
     cpfile = file
+    
     #copy file to current directory
     os.system(f'cp "{cpfile}" .')
 
-
+    
     newfile = file.split('/')[-1]
     os.system(f'mv "{newfile}" "{newfile.replace(" ", "_")}"')
     newfile = newfile.replace(" ", "_")
     cpfile = cpfile.replace(" ", "_")
     name = newfile.replace('.ipa', '').replace('_', ' ').replace("decrypted","Decrypted")
     identifier = re.sub(r'[^a-zA-Z0-9-]', '', name.replace('.', '-').replace(' ', '-')).lower() + '-ipa'
-
     # Extract iPA
     zipfile = newfile.replace(".ipa", ".zip")#.replace(" ", "\ ").replace("(", "\(").replace(")", "\)")
     copyfile(newfile, zipfile)
-    os.system(f'unzip "{zipfile}" > /dev/null')
-
+    os.system(f'unzip -o "{zipfile}" > /dev/null')
     # get only directory in Payload folder to find Info.plist
     infoplist = os.listdir("Payload")[0] + "/Info.plist"
 
     bundleid = ""
     iconfile = ""
+    appname = ""
+    version = ""
 
     # get bundle id and icon file name from Info.plist
     with open(f"Payload/{infoplist}", "r") as f:
-        contents = f.read()
-        soup = BeautifulSoup(contents, 'xml')
-        bundleid = soup.find("key", string="CFBundleIdentifier").find_next("string").text
-        iconfile = soup.find("key", string="CFBundleIcons").find_next("array").find("string").text
+        try:
+            contents = f.read()
+            soup = BeautifulSoup(contents, 'xml')
+            bundleid = soup.find("key", string="CFBundleIdentifier").find_next("string").text
+            iconfile = soup.find("key", string="CFBundleIcons").find_next("array").find("string").text
+            appname = soup.find("key", string="CFBundleDusplayName").find_next("string").text
+            version = soup.find("key", string="CFBundleVersion").find_next("string").text
+        except:
+            contents = readPlist(f"Payload/{infoplist}")
+            bundleid = contents["CFBundleIdentifier"]
+            appname = contents["CFBundleDisplayName"]
+            version = contents["CFBundleVersion"]
+            try:
+                iconfile = contents["CFBundleIcons"]["CFBundlePrimaryIcon"]["CFBundleIconFiles"][0]
+            except:
+                try:
+                    iconfile = contents["Icon files"][0]
+                except:
+                    try:
+                        iconfile = contents["CFBundleIconFiles"][0]
+                    except:
+                        # check if Icon.png exists
+                        if (os.path.exists(f"Payload/{os.listdir('Payload')[0]}/Icon@2x.png")):
+                            iconfile = "Icon@2x.png"
+                        elif (os.path.exists(f"Payload/{os.listdir('Payload')[0]}/Icon.png")):
+                            iconfile = "Icon.png"
+                        else:
+                            raise Exception("Icon file not found in Info.plist")
+                            
+                
 
+    # if name contains more than 2 periods, replace name with appname
+    if (name.count(".") > 2):
+        print("Looks like we failed to parse the name from the filename. Using the app display name in the Bundle ID instead")
+        name = appname + " " + version + " Decrypted"
 
     # find icon file in full inside Payload/*/, since iconfile is only beginning of file name
     iconfile = [f for f in os.listdir(f"Payload/{os.listdir('Payload')[0]}") if iconfile in f][0]
@@ -87,8 +119,13 @@ for file in filelist:
         os.system(f'cp "Payload/{os.listdir("Payload")[0]}/{iconfile}" . > /dev/null')
     else:
         os.system(f'cp "Payload/{os.listdir("Payload")[0]}/{iconfile}*" . > /dev/null')
+
+
+    if (testmode):
+        input("Test mode, holding Payload folder open. Enter to continue...")
     os.system(f'rm -rf Payload')
 
+    
     # Archive.org cannot read the icon pngs, so a specific thumbnail is needed
     os.system(f'sips -s format jpeg "{iconfile}" --out _itemimage.jpg > /dev/null')
